kexec: add support for adding more ACPI tables
In some cases we want to extend the ACPI tables with more.

This change adds a -a switch to kexec, which allows one to add
a single table. For now that is all we need; if we want more we
change the operation later, once we are sure this is the usage
mode we need.

Note that due to Linux limitations, the -a switch only
works with multiboot kernels.

We've extended the multiboot package to add an ACPI operation
on multiboot. We'll probably be moving most of the code over, and
just hand the ACPI method path name. There's no reason to leave
that junk in the kexec command. But I'd like a little review
of the general idea.

This is tested and seems to work with our toy multiboot kernel.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>