multiboot: add ACPITable function
For esxiboot or other programs that will create a Tabler,
we need a function to take a tabler, not a pathname.

Another nice result is that we can now add any number of tables
to the ACPI table for multiboot. We can even add 0 tables and just
force a remarshaling if we so desire.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>