u-root: automagically adjust absolute paths with a warning
u-root will not create archives with absolute paths in them.
It's too dangerous: one wrong command in a wrong window and
your machine is toast. I've done it.

The path rewrite commands of the form -files /etc:/etc
were therefore forbidden, but people find this inconvenient.

Now, when we have an absolute path as the destination, we will
adjust it to be relative to /.

A test is included and a test in initramfs is adjusted.

One problem with our testing package is that the warning messsage
appears on stdout when you run the test. We might want to fix this.
I'd like to keep the warning but maybe we don't want it?

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>