cpio: GetRecord:ignore Access errors; WriteRecord: ignore file read errors
If a file is not accessible, GetRecord should return an error
and let the caller figure out what to do. The caller can ignore it
as the cpio command does.

In WriteRecord, any read errors should be ignored
as GNU tar and cpio do. They are interpreted as a short
read. Further, in newc format, a short read should
be extended to the reported FileSize.

This fixes bugs when using cpio on synthetic file systems,
so we match behavior of (e.g.) GNU cpio and tar.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>