Deleted duplicate parsers for /proc/filesystems
Deleted the fstypes() function from ../../diskboot/device.go file
and instead, imported the function GetSupportedFilesystems() from
../../storage/filesystem.go in the device.go file. Both returned
ALL block file systems supported by the linuxboot kernel.

The last file, ../../util/root_linux.go, implemented the parsers
for /proc/filesystems differently. First, the function took in a
file system the user is specifically looking for and checked through all of
the block file systems supported by the linuxboot kernel to see if the one
the user is looking for exists or not. If it doesn't exist, it returns an
error message.

Signed-off-by: Urvisha Patel <patelvisha2007@gmail.com>