Previous Commit: Two test scenarios: close & exit or kill & exit
Kill yes command is unusual. More common implementation: to
exit when the pipe it writes to closes. Implement two scenario
to test yes command: close stdout pipe and make sure yes exists
when closeStdout bool is true, and kill yes when closeStdout bool
is false to make sure two implementation work.

2nd most recent Commit: Fix minor changes

Delete the line "v.closeStdout=false" line after sending the error
message in regards to closing standard out pipe as it serves no purpose.
Creating 6 tests where half are true and half are false, testing with
the same output allows us to skip setting the bool to false. Fix the message
"Complete the child process: got %v, want "signal: broken pipe"
because the error will not provide a nil, and because we are forcefully
closing the pipe, it will print the message broken pipe, letting us know
the pipe is closed and we received the expected output.

Also, check if the stdOut pipe provides a nil error. If it does, the pipe did
not close the expected way.

Most recent commit: Delete the 'else' nil for closing pipe

Delete the 'else' clause, which checks for err == nil.
It is complex and unnecessary for this specific type of testing.

Signed-off-by: Urvisha Patel <urvishapatel@google.com>