provide a function to get net interfaces from an RE
It's handy to be able to take an RE and get a set of interfaces
from it. We've found a need for this function in more than one
program now; time to make it part of a package.

The Interfaces function takes an RE and returns a list of
netlink.Link that matches it, or an error. It is an error
for the returned list to be empty.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>