abi: move to strace/internal
pkg/abi leaked from gVisor in the copy of strace. However, it should
largely be replaced with golang.org/x/sys/unix where we can, when
someone has the chance.

In gVisor, pkg/abi makes sense because you may be stracing a Linux
application on an arbitrary $GOOS host, so x/sys/unix will contain
$GOOS's stuff, not Linux stuff.  In u-root, the tracer's OS is
definitely the same as the tracee's.

Signed-off-by: Christopher Koch <c@chrisko.ch>