WIP: use the hugelgupf/9p package for a builtin 9p server
cpu will now by default use a builtin 9p server.

You still have the option of using an external program,
which is very useful for debugging.

Note this does not work at present b/c the ssh channel is
not backed by a syscall.Conn and the 9p package expects that it is.

It will, again, still work with an external 9p server.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>