Do not include the cmds/boot/uinit in the "all" template
When we merged systemboot in we included the systemboot uinit
in all by accident. Explicitly select the commands we want in
boot.

Also, use the power of the glob for the cmds/boot to include
cmds/boot/*boot*, which includes any new ones we add and nicely
excludes the uinit.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>