dhclient: remove broken Classless Static Route option
Rolling back 'cause it's broken, and I don't have the time to fix it,
but I also need it to not be in the way for an internal project.

It works for a /32 IP, but any non-/32 IP with the following classless
static routes fails to have routes configured correctly:

Subnet Mask: ffffff00
Router: 10.0.0.1
Classless Static Route:
  route to 10.0.0.1/32 via 0.0.0.0; route to 0.0.0.0/0 via 10.0.0.1

(Rejected by RTNETLINK.)

They're so simple, yet so hard.

Signed-off-by: Chris Koch <chrisko@google.com>