stboot: pkg/bootconfig
Refactor stbooy such that in only uses iso images instead of kernel +
initramfs (to overcome the problem, that we need to include the
filesysten in  the initramfs - or find another way to load it after
kexec). Also refactored the code to use mor functions and do not have a
huge main() function.

Missing:
* Right now we abusing the kernel field in the manifest to save the path
of the iso image. This need to be adjustetd.
* No tooling to generate the iso image needed for stboot

Signed-off-by: Christian Walter <christian.walter@9elements.com>
Signed-off-by: Jens Drenhaus <jens.drenhaus@9elements.com>