Package for manipulating Memory Technology Devices
Realistically, nowadays, that means "spi", but there can
be more to it.

This code is designed to compress efficiently for embedding
in FLASH parts (similar to what we did for the pci package)
and, given that the tables are basically one pass, we don't
build any maps as we might have.

The test passes on linux with an mtd test device.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>