localboot: Unquote GRUB strings unconditionally
This works around issues where some distros or configurations use the
legacy GRUB config filename, grub.cfg, instead of something more obvious
like grub2.cfg. It stops short of removing the version logic all
together though since it acts as a test to see if we're dealing with a
GRUB config at all (this might be further simplified, if desired).

Some Linux kernel command-line parameters use special characters, for
example memmap=128M$1G. The $ can be escaped (or quoted) in GRUB2 using
backslash, i.e. memmap=128M\$1G, which is why we have unquoting logic in
grub.go.

The issue we encountered is that a distro installed a GRUB2 config file
as grub.cfg, with memmap=128M\$1G on the Linux command-line.
Consequently GRUB legacy was detected and unquoting was not performed,
so the parameter was passed into the kernel with the backslash, making
it invalid and causing other issues.

GRUB legacy strings should not contain GRUB2 quoting so it should be
safe to run them thru the same unquoting logic.

Signed-off-by: David Hendricks <david.hendricks@gmail.com>