fix cpu to use the -v not -verbose with dhclient
at some point this changed.

While we are at it, fix the usage. specifying the host follows the ssh model,
so the args are now:
cpu host [options]
and the -h switch is gone.

Add a usage function. I think this usage may be what we should
do in general; write usage to a bytes.Buffer, then print it all
out at one time.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>