Remove annoying rc.elv error message on startup
elvish would print an error message on startup if rc.elv
did not exist.

Turns out this was an error in the original; the author
expected filepath.Abs to return an error if the file
did not exist, and that is not the job of Abs.

Move the error test to the right place; no more annoying
error message. This has been a surprisingly common source
of confusion over the years so it's nice to have it gone.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>