RFC proposal
We need an RFC mechanism so here is one idea.

I'm pretty adamant on Latex. We tried markdown for the book
and it did not end well in my view. With Latex we can get
good looking output with good support for automated
figure numbers, citations, image inclusion, and so on.
Latex is a standard for publication nowadays.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>