scuzz: a package for controlling disk devices
The scuzz package defines Request and Disk interfaces.

Requests are requests to operate on Disks.

The Disk interface defines functions to create Requests
and to invoke an operation, given a request.

We provide an implementation that should work for a Linux
SCSI Generic device. We include a test which should ensure
the packets are correct. The packet header and data in this
case are generated from a fork of hdparm maintained at
github.com/rminnich/hdparm.

We also include a sample hdparm command.

We currently only support one request, namely, Unlock.
We expect to need more :-)

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>