scuzz: add ATA identify command
The ATA identify reads information from those disks
which support the command.

This PR adds the Response interface, which allows programs
to query the result.

We separate what look like transport errors, i.e. the system
call failed for some reason, from target errors, i.e. the system
call worked but the drive rejected the command. The kernel does
not make such a seperation, which calls our seperation into
question.

Error determination from the status block is a work in progress,
not expected to be completed quickly.

This is nevertheless a basic starting point for an identify.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>