New cpu command, dep update for it.
This comtains the internal 9p server from gvisor. It has numerous small fixes.

The client now writes a nonce to the server environment, and requires
the server to write the nonce back on the ssh tunnel before it will accept
the mount.

The client now only accepts one connection.

The doc now includes an advisory on when it is safe to use cpu and when
it is not, as well as an explanation of how the nonce is used.

We add the gvisor license header to checklicenses

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>