Adding integration test orchestrator
Problem: we needed to be able to pass a custom initramfs to a test instead of having the test use the generic initramfs.

I couldn’t find a better way of passing the initramfs to a test besides passing it as a flag. But that gets annoying if you  want to run all tests at once, so I wanted to have a test orchestrator that made it easy to run all the tests at once, which also let you specify specific initramfs.

This led to eventually restructuring the way integration/ is set up. The
modified integration/README.md explains the new structure.

It's a little more complicated than earlier but it should still be just
as easy to run tests, and hopefully gives us more flexibility in
writing/running tests.

Comments on how to improve this are very welcome!

Signed-off-by: pallaud <plaud@google.com>