Modify tests
The goal is to have most tests use the generic initramfs and test
commands method unless a custom uinit is really needed.

The kexec case is definitely not the most elegant, but is at least
a POC that we can test it using the generic uinit.

I also split up uinit_test.go and moved the testdata dir to be in
generic-tests/.

Signed-off-by: pallaud <plaud@google.com>