pox: trim PWD from paths in the archive
pox uses the full path of files, including files you passed to pox and
any shared libraries used by those files.  If you have a program
installed somewhere non-standard by configuring with --prefix, pox would
include all of the directory paths in the archive.

This commit trims the directory where you are running pox from the path
of any files.  A simple usage then is to cd into the directory you want
to pox, then run:

	$ pox -c `find . -type f`

Your program would then appear in the chroot in the same way as it would
if you had installed it to e.g. /usr/local/ instead of your prefix.

This makes it easier for pox to find your binary, since it is likely in
bin/, and a user can extract a pox file directly into /usr/local/.

Signed-off-by: Barret Rhoden <brho@google.com>