pox: add an option to use zip instead of squashfs
CONFIG_SQUASHFS and /dev/loop-control are not universally available in
Linux, which makes it difficult to run pox files on other systems.
unzip is more readily available.

This commit adds '-z': zip.  Instead of making a squashfs file, -z will
make a zip file.  When running the pox, use -z again to unzip the pox to
a tmpfs directory.

Signed-off-by: Barret Rhoden <brho@google.com>