u-root: default to build=bb
When we started u-root in 2011, on Go 1.2, the toolchain was
based on Ken's Plan 9 C compiler toolchain (in fact, the amd64
toolchain was derived from the DOE-funded C toolchain created
by Vita Nuova for DOE Fast-OS).

That compiler was so fast that I ran, by mistake, with commands always
being compiled and it was months before I noticed it.

Time flies. Go is not as fast as it was. Commands take many seconds to compile.
It's not really viable to have a Go compiler in FLASH at this point.
Even little disks are very big, so the space argument no longer holds for root
file systems. Almost no one (and maybe no one) uses source mode any more.

Make the default be bb, which everyone uses.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>