Get completion rules right
If you run
INTERACTIVE=1 go test -run TestSh

you can try what I think may be a good enough completer.

The first arg on the line will be determined from $PATH;
then we switch to a file completer.

ash is also modified to work with this. It seems to work
fine SAVE there is not  prompt yet; WIP.

Try this if you can and see if it is close.

If it works,  we have an advantage:
- much smaller code and binary: 2.7M vs. 5.1M for elvish
- code is easier (for me) to understand
- completer package is useful in other programs

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>