securelaunch: helpers: Add mountCache to speed up boot
Currently we are using sluinit to do things in the following sequence:
1. get tpm handle
2. get policy (mount all devices until we find policy file)
3. collect evidence (DMI, Files, Storage, CPUID) (all collectors mount devices except DMI collector)
4. measure target kernel & initrd (mount device twice once for kernel and once for initrd)
5. dump eventlogs to disk (mount device for target location of dump)
6. Launcher for boot.(mount device for getting kernel and initrd files)

Notice how many times we are mounting and unmounting disks...
Now lets take the case when the user enters same disk for all these (e.g sda2).
That would mean in one sluinit sequence, we will mount and unmount sda2 about 9 times

Sigh... That really is inefficient consider a single mount attempts to match different kinds of
filesystems(ext3, ext4, xfs) before finally mounting....a single mount typically can take around 15-20 secs.

My proposal to fix this inefficiency is to have a mount cache.
first time we mount a device, keep the device mounted and store the mountPath in a map(cache).
that way if a device is already mounted...we can use the map to lookup for a given device and get the mountPath.
At the last step of sluinit before kexecing, we go over the map and unmount all keys left in the map.

We would be needing a lock for cache since maps in go are not concurrent safe.

This also makes GetMountedFilePath function simpler since we no longer need to return mountPath.

Note that mount flags make things slightly complicated. if a device is in cache with read flag
and we want to write to the same device, we would need to unmount the device, delete the map entry
and mount the same device again with write flags. Our goal should be to minimize number of remounts of
same device.

Signed-off-by: Simran Singh <simran.p.singh@oracle.com>