OSX/pxeserver: Change response for DHCP to match subnet
DHCP packets being sent to 255.255.255.255 are being
swallowed whole. In a lot of ways, 255.255.255.255 is
a very odd place to send: it technically matches all
interfaces, but we're sending on one interface.

My brain seems to recall that you can send a broadcast
packet on a subnet, given the subnet address, but you
can't send to the broadcast address. But that's probably
just a Total Recall.

This fixes the packet swallowing problem and the Atomic
Pi can once again be booted from pxeserver running on OSX.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>