madeye: create Multi Architecture Disk Images
This is an original goal of u-root: that we could build a root file system
and use it on different architectures.

This is called MADI: Multi Architecture Disk Image. It has binaries for several
architectures, rooted at, e.g., /amd64, /arm, and so on; with binaries in, e.g.,
/amd64/bbin, and init in /amd64/init.
To boot on, e.g., x86_64, you set init=/amd64/init

It's spelled madeye so people know how to say and, because, some might think it mad.

This idea is not original with u-root; it's how Plan 9 works. It was the reason for the
"source mode" in original u-root: four binaries only, the rest source. Busybox mode
did not come until 2015 or so, and it turns out it lets us do the same thing with
one binary or five; we can have both busybox and source mode MADIs.

Source mode and busybox mode can share a single root file system, so I'm starting out
supporting just busybox mode.

Rather than complicate the u-root command with a bunch of options for multiple architectures,
we take the approach here of implementing a tool that concatenates u-root cpio's, which
is neat as it allows after-the-fact construction of MADI. Names are translated as needed,
as are symlink targets.

This is a WIP for a bit longer, but the driver was the class we're putting together, where
we want lightweight pxe servers and we don't much care about architecture. madeye will let
us construct USB sticks that boot on atomic pi or raspberry pi, with a very simple root
file system.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>