pox: just use the list of names retured by ldd.Ldd()
ldd.Ldd returns ALL the names, including the names of
the input files. We appended the ldd.Ldd results
to the file names, resulting in us adding the binary
names twice.

This has not been a problem in several years, until we hit
a weird case today where adding a file twice failed.

Just use the ldd.Ldd() list.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>