msr: move definitions to package, add Test function, add -verify to lockmsr
There were constants defined in lockmsr that are
useful to have in the msr package; move them. Place the
Intel-specific MSR settings in an array.
Add a Test function to verify settings.

Add a Verify function to the package to make writing
tests easy. Verify takes a vendor string which must, for
now, be GenuineIntel.

Add test code, now that constants and a Test function
are available.

Add -verify option to lockmsr.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>