strace: remove special execve syscall-stop logic
It turns out that when we first stop at execve, it's not actually a
syscall-exit-stop as I first thought. (The arguments were never filled
out correctly in the printed versions, and then I noticed that all the
call registers are 0.)

We eliminate printing this execve and the arguments for now.

Regular strace shows an execve exit at the beginning, implemented
through other means:

    execve("../../../pkg/strace/test/hello", ["../../../pkg/strace/test/hello"], 0x7fff05a7ff30 /* 46 vars */) = 0
    brk(NULL)                               = 0x555556f29000

whereas for now we will start with the break:

    [pid 181215] E brk(0x0)
    [pid 181215] X brk(0x0) = 0x555555cf1000 (176.28µs)

This commit explains, in code, why and what our options are later.

Signed-off-by: Chris Koch <chrisko@google.com>