Fix issues in boot2
When trying to boot RHEL 7.8, we noticed that parsing the GRUB
config failed. This is because diskboot/parser.go expected to find
the exact string "linux" and "initrd", and the RHEL config had
"linux16" and "initrd16". The changes here make it so we can
parse that correctly, and make debug messages a little more informative.

Signed-off-by: pallaud <plaud@google.com>