cmds: don't call flag.Parse in init
When flag.Parse is in main, it is _only_ called in the command binary.

When flag.Parse is in init, it is _also_ called in the test binary. The
test binary, however, likes to register its own flags, which due to init
ordering usually happens after the command's inits.

In Go 1.13, flag registration was delayed to be later in the init
process, which is how this issue came up.

See also https://github.com/golang/go/issues/31859

Signed-off-by: Chris Koch <chrisko@google.com>