export GoTPMAlg()
on each measurment, securelaunch package sends pcr events to kernel
via sysfs. these events are marshalled in the order "txtlog" package can parse.
one of the variables in this marshalling sequence is the hash algorithm
used for the measurement and is of type tpm2.Algorithm. Exporting GoTPMAlg
allows securelaunch package to easily access this type.

Signed-off-by: Simran Singh <simran.p.singh@oracle.com>