cmos: add non-constrained docs file
cmds/core/io/cmos.go is only built into the io package if it's amd64 or
386. In the Go build system, that means pkg/cmos is never even compiled.

However, in blaze/bazel, all dependencies are always built, used or not,
since it doesn't know any better. That means pkg/cmos has to be
buildable on other architectures than amd64 or 386.

If you just add the build tags "+build amd64 386", pkg/cmos won't be
compilable by blaze/bazel because it's empty.

Hence, we need a neutral file in there that always builds.

Signed-off-by: Chris Koch <chrisko@google.com>