AMD System Management Network dumper
This tool allows probing of the System Management Network.

It allows reading many words at a time, but only setting one
word at a time on a northbridge. However the northbridges are specified
with a glob, as in the pci command, so one can dump all values of a given
range of SMN for several northbridges, e.g. one has a system accessible
via the cpu command, we can see parts of SMN as so:

cpu 10.0.1.18 ./smn -s 0000:*:00.0 13b102f0

If we want to further restrict it, and add more SMN addresses:
cpu 10.0.1.18 ./smn  -s 0000:[02468]0:00.0 13{b,c,d,e}102f0

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>