menu: separate Do into Load and Exec
For some uses cases, we need to clean up the OS before we kexec.
Ideally, we do that by transferring execution back to the caller and
letting them call kexec, so they can do what needs to be done.

However, we also want to give users the choice to re-try or pick another
OS to boot if e.g. a kernel file download fails or the kernel executable
corrupted.

So Load becomes a re-tryable action in ShowMenuAndBoot, but Exec is
reserved for the caller to call after clean-up.

Signed-off-by: Chris Koch <chrisko@google.com>