syslinux,curl: support relative file paths correctly
Previously, we were generating invalid URLs in our custom parseURL
functions by using relative paths.

Our paths would be something like file://foobar/boot.cfg, which would
parse into

&url.URL{
  Scheme: "file",
  Host: "foobar",
  Path: "/boot.cfg",
}

But what we actually want is for it to parse into

&url.URL{
  Scheme: "file",
  Host: "",
  Path: "foobar/boot.cfg",
}

However, there is no valid representation of that because it's
ambiguous. The only valid thing would be where Path =
"/foobar/boot.cfg".

So, to trick the URL parser, we use Host = "." as a custom indicator.

Signed-off-by: Chris Koch <chrisko@google.com>