dhclient: wait on iface OperUp/OperUnknown instead of FlagUp
We used to wait on FlagUp, which actually gets set the moment you call
LinkSetUp, and has no bearing on whether the link is actually
operational. Sometimes the nic is still waiting for carrier and we
incorrectly proceed while the nic is still down. Also take linkUpTimeout
as an argument instead of hardcoding it in the package.

Signed-off-by: Gan Shun Lim <ganshun@gmail.com>