exp/pxeserver: support DHCP IP provisioning from /etc/hosts
pxeserver has proven to be valuable for provisioning netboot clients.
So valuable, in fact, that the need to support more than one host
has arisen.

This change lets pxeserver support an arbitrary number of clients,
on any active NIC, in IP4 for now, by using /etc/hosts to hold MAC to IP mappings.

No special file formats are needed. Users will need the ability to
edit /etc/hosts; but if users are running any kind of pxe or dhcp server,
you've got that already.

When pxeserver receives a request in the dserver4 loop, it will use the
net.LookupIP for a hostname  of the form: 'uMAC', where mac is the stringified
version of net.HardwareAddr, and u is a single-leter prefix to make it a hostname,
not a number. LookupIP returns a []net.IP. The code
uses the 1st element, converted to an ip4 with To4(), as the IP address.

For legacy use, if the lookup fails, and if pxeserver has not responded to an unknown
host before, it returns the default address specified by the your-ip flag.
If the default is already in use,pxeserver prints an error message of the form:

2020/07/13 13:55:08 Not responding to DHCP request for mac 00:07:32:4c:25:36, which does not match 00:07:32:4c:04:e6
2020/07/13 13:55:08 You can create a host entry of the form 'a.b.c.d [names] u00:07:32:4c:25:36' 'ip6addr [names] u00:07:32:4c:25:36'if you wish

which can be used to create a new hosts file entry.

The lookup does not choose to use /etc/hosts; although the current use is to add mappings to /etc/hosts, they can
be added to any source of hostname-to-IP address data sources.

The lookup is stateless; pxeserver does not cache, but rather does a new lookup each time, so that
should the host database, the new information can be used.

Example: one might have a network of several Atomic PIs. One can bring them up one at a time, and look
for messages of the form:
2020/07/13 13:55:08 Not responding to DHCP request for mac 00:07:32:4c:25:36, which does not match 00:07:32:4c:04:e6
2020/07/13 13:55:08 You can create a host entry of the form 'a.b.c.d [names] u00:07:32:4c:25:36' 'ip6addr [names] u00:07:32:4c:25:36'if you wish

At that point, one might add this line to /etc/hosts:
192.168.0.4 api3 u00:07:32:4c:04:e6

On the next dhcp request, api3 will be provided with an ip.

This approach has many benefits: the file format is simple; leases need never expire; and, most important, once one assigns
a MAC to IP mapping, it is available for all other programs (e.g. cpu), which has proven very handy in practice.

On a recent experience setting up a small network of atomic pis, the task was accomplished with far less effort, and in
far less time, than was possible with isc-dhcpd and its companion tools.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>