ns: bind, mount and unmout for plan9
plan9 has a file format for building name spaces

	http://9p.io/magic/man2html/6/namespace

main functions for all there apps are identical, with the exception
of their documentation. This is a happy side effect of the fact that
the description files are in the same format their command line
counterparts, thus the file parser is used to parse the os.Args too.

Signed-off-by: sevki <s@sevki.org>