trampoline: find info,entry,magic location using funcPC
The crux of the problem is that we're trying to find these symbols'
location in order to insert some bytes at each. Until now, we've done
this by putting some strings before the symbol, and finding the string
in the trampoline, and then replacing the bytes in the function _after_
assuming a Go compiler's alignment of 16 bytes.

Go 1.15's
https://github.com/golang/go/commit/30f8074e3510da8c39f879cfdbde600eb876a79e
changes the default alignment of functions from 16 to 32 bytes.

We could decide to use 16 bytes for <= Go 1.14 and 32 bytes for >= Go
1.15. Instead, each symbol is treated like a function so we can use funcPC
to find their location. This fix works in Go 1.13, 1.14, as well as 1.15
without keying off an internal Go compiler alignment value.

Fixes #1803

Signed-off-by: Chris Koch <chrisko@google.com>