kexec: support gzip-ELF files so ARM64 works
ARM64 kexec_file_load only processes ELF files.

But on many ARM64 Linux distros, the ELF file is
gzip'ed, with names like vmlinuz-x.y.z-whatever.

Move the tryGzipFilter code to boot/util, and
make it exported as TryGzipFilter, so that it
can be used in kexec as well as multiboot.

Not tested on ARM64. Will need to write a test if
this approach is acceptable.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>