grub: Add support for the search command
This adds support for the grub `search` command as defined by the grub
manual. This allows the kernel and initrd to be loaded from a file
system other from the boot drive. CentOS exercises this command and has
been added as a unit test.

The following options are included:

* `search --file`: Set root to the file system which contains the given
  file.
* `search --fs-label`: Set root to the file system with the given label.
* `search --fs-uuid`: Set root to the file system with the given UUID.

The aliases `search.file`, search.fs_label` and `search.fs_uuid` are
also supported.

Signed-off-by: Ryan O'Leary <ryanoleary@google.com>