cpio on plan9: convert symlinks to binds
Plan 9 does not have symlinks.

"Symlinks are a botch" -- David Korn

"But they're so damn useful" -- everyone else

Anyway, for now, if cpio sees a symlink, it will now do a
bind by creating an empty file and binding it to the symlink
target (a.k.a. "contents of the cpio record").

It's the best that can be done at present. It is only
useful when unpacking a cpio into a ramfs. But Harvey
needs that.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>