Modify the bb main so bb can be a #! interpreter
See the main.go for a long discussion.

Short form: we need a way to run on file systems without
symbolic links.

This change lets bb run as the interpreter in a #! file.
Due to differences in kernels, we mark the argument with a #!.

So, e.g., /bbin/ls on Plan 9 or a VFAT file system becomes:

Tested on Linux and Plan 9 via the following script:
192.168.0.7# cat /tmp/ls
192.168.0.7# /tmp/ls
acme.dump
bbin
...

The next change will be to have u-root optionally generate
these files instead of symlinks. Note that it will be required for
Plan 9 but will work on any Unix system as well.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>