block: Add 'p' before the partition number when necessary
localboot -guid cannot boot because the device name in the partition
map key doesn't have 'p' before the partition number, so it cannot find
the block device from the map. Add 'p' if the last character of the
device name is a digit. For instance, nvme0n12 would be nvme0n1p2.

Signed-off-by: Johnny Lin <johnny_lin@wiwynn.com>