pkg/iomem,cmds/core/io: Add support for x86 CPU IO instructions
Sometimes going via /dev/port is too slow.

Add support for x86 raw assembly IO. Add an init function that
does an iopl(3). This function need not succeed; it sets an
error value so that if an attempt is made to use those instructions,
and the iopl failed, the iopl error will be returned.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>